[feat] HistoryManager 클래스 추가 및 디렉토리 수정 #61
History에 대한 로직을 담당하는 클래스
